requirements.txt
aiohttp>=3.10.0
beautifulsoup4>=4.12.2
python-telegram-bot>=20.7
PyYAML>=6.0.1
lxml>=5.0.0
python-dotenv>=1.0.0
